Created an external library loader to support functionality introduced in the IO_Handler class (forgot to add this as part of that commit).
This just returns the last folder name in /Libraries/External/<libraryname>/
which should be the version number or "current" for git submodules. For example,
if you have smarty loaded, this may be /Libraries/External/Smarty/3.1.11/ which
would return "3.1.11", or for php-openid, we'd get "current" from
/Libraries/External/PHP-OpenID/current.

I have added unit testing, and some unused, but required directories to support
that unit testing.